automation/requirements.txt
pyyaml
munch
requests